Add unit tests for remark/rehype plugins
See #17

Migrate unit tests for our custom remark and rehype plugins from the
archived `gravitational/docs` repo. Convert these tests to Jest, which
unlike the uvu testing tool we used for `gravitational/docs`, is still
being maintained. Add a GitHub Actions workflow to run the tests on pull
requests.

To get the `remark-includes` test to pass, edit the `updatePaths`
function to only extract a version number from the path of a partial if
the partial falls under a directory path for production content.