Edit the absolute link linter
Allow JavaScript code as the value of the `href` attribute. This allows
us to import assets as MDX modules so we can refer to the asset paths in
HTML tags. Otherwise, the linter encounters a runtime exception and
fails even if there is a comment that ignores messages from the linter.

An example of the linter failing is gravitational/teleport#51193, which
imports a bot avatar logo as a module so users can download it.