Allow submodule version aliases for import paths
Closes #195

Add a `remark` plugin that lets a docs author include an `import`
statement that points to a directory in the current version submodule of
the docs site. This makes it possible to import the path of a static
asset or, with the `!!raw-loader!` directive, the full content, using an
`import` statement, without having to know which versioned submodule an
asset is in at a given time.

To use this, add the string `@version` to the start of an import path,
e.g.,:

```jsx
import PNGPath from '@version/docs/img/myimg.png'
```

In a `gravitational/teleport` clone at the current default version of
the docs site, the new plugin would add the following:

```jsx
import PNGPath from '@site/content/17.x/docs/img/myimg.png'
```

Note that we currently use a workaround in which we place some static
assets in the `/static` directory, which exists outside the submodule
directory tree. This is approach is cumbersome since it requires a
change to `docs-website` for every asset we want to include using an
`import` statement.

The `@site` alias is a Docusaurus feature that the docs engine replaces
with the root path of the Docusaurus project. The plugin fills in the
path of the current `gravitational/teleport` submodule.

To import the content of a text asset, rather than the file path, you
would use the `!!raw-loader` syntax as you would for any Docusaurus
asset:

```jsx
import PNGPath from '!!raw-loader!@version/docs/img/myimg.png'
```