Fix dependency configurations
Revert unnecessary dependency changes. Remove @types/estree. Also remove
an extraneous `mdast-util-mdxjs-esm` mention in `package.json`.

Add swc/core to the list of dependencies with allowed licenses for the
dependency review workflow. While the package uses the Apache 2.0
license, which is already configured as allowed in the workflow, the
workflow also detects an unknown license that does not appear to be
present in the package's code repository.