Add a custom 404 component
Closes #121
Closes #64

Replace the default Docusaurus 404 component with one that is more
helpful to docs users.

Add the following links to the 404 page so a user can find content
similar to what they are looking for:

- A link to the nearest category index page, reducing the likelihood
  that a user needs to navigate through the docs sidebar tree manually.
  For example, a visit to `/docs/enroll-resources/blah` would result in
  a link to `/docs/enroll-resources/`.
- A link to any pages with the same slug as the 404ing page, in case we
  moved the page but did not provide a redirect.
- A link to the main docs home page.

This change also adds a search bar so the user can use what is often the
most helpful tool for navigating to Teleport docs content.

Finally, this change adds a link to create a GitHub issue with the
404ing path so we can rectify any problems on our end.