Fix bug in rehype-hljs
Closes #218

In the predecessor to the `rehype-hljs` plugin within the
`gravitational/docs` repo, gravitational/docs#506 addressed an
unanticipated aspect of the HTML AST within a code snippet by branching
between different kinds of nodes.

However, this branching was not applied throughout the plugin, meaning
that in some cases, the plugin mistakenly removes the `code` element that
is the child of a `pre` element.

The copy button in our swizzled Docusaurus `Pre` component copies text
in a code snippet by querying the `.hljs` class, which our syntax
highlighting library adds to `code` elements. As a result, in cases
where the plugin removes the `code` element, it makes the copy button
nonfunctional.

This change edits the `rehype-hljs` plugin to branch on attributes of a
node when modifying the final AST--a step neglected in
gravitational/docs#506--avoiding the `code` removals. It also stops the
plugin from adding the `hljs` class to `span`s, since this class is only
supposed to modify `code` tags. This prevents the copy button from
copying snippet values twice.