Fix TabItem `value` assignment
Fixes #200

The current `TabItem` wrapper uses `nanoid` to assign a random value to
the `value` property, which is what we did at the AST level prior to
the changes in #183. This change uses `useId` instead, which is the
idiomatic React technique for achieving a similar result. This approach
also avoids the issue in #200.