Lint required frontmatter fields
Require a `product` field and `type` field. We will use this to ensure
that every docs page specifies the Teleport product it belongs to (e.g.,
Identity Governance) and the kind of guide it is (e.g., how-to guide).
We plan to use these fields to create custom views of pages in the docs
that match a particular category or combination of categories.