Emit gtag events on code copy button clicks
Help understand how documentation readers use docs pages to perform
operations on a Teleport cluster by adding an anonymous gtag event when
a user clicks a code snippet copy button. This includes both the large
button that copies an entire multi-line snippet and the smaller button
that appears next to a single command within a `code` snippet.

To begin with, gtag event payloads are minimal and contain no
information about the content of the snippet:
- `label`: the language of the code snippet, e.g., `code` or `yaml`
- `scope`: whether the copy button is for a single line or the whole
  snippet
- Default fields such as the path of the page.

As part of this work, include an optional `gtag` prop in components that
emit a gtag so we can easily and explicitly assert against gtag events
in tests.