Refactor gtag calls to use `trackEvent`
Use the `trackEvent` function added in #250 to send gtag events. Make sure we
are using a single function to send these events in order to keep our
monitoring consistent. This change adds to `trackEvent` in order to
allow for mock `gtag` functions in testing. If there is no `gtag`
function available, `trackEvent` logs the event to the console.

Also move code from `src/utils/gtag.ts` to `src/utils/analytics.ts` so
all analytics-related code is in one place.