Record index and count in code copy button events
Indicate how far down a page a `Command` line or `Pre` snippet is when a
user clicks a code copy button, including both the whole-snippet button
and the individual line button.

Add a `PositionProvider` context provider. Components register
themselves with the provider when rendering so the parent component
doesn't need to know about its children in advance. This is necessary
for docs pages, since a `DocItem` doesn't know in advance of MDX parsing
how many `Pre`s are in the page. Similarly, the `Snippet` component
results from the `remark-hljs` plugin and doesn't know how many
`Command`s are in a code snippet.

Keep the total count of elements the same between renders by using
`useRef` and `nanoid` to give each `Pre` and `Command` a unique ID that,
unlike the result of `useID`, stays the same between renders.

To track the position of each `Command` within a `Pre`, have `Pre` wrap
its `Command`s in a nested `PositionProvider`. Pass a callback to
`PositionProvider` so it can report the total number of outer `Pre`s, as
well as the position of each `Pre`.