Convert two Vale structure linters to remark
See #204

This change turns two Vale linters into a `remark` linters. This way,
docs pages can add a comment directive to ignore the linter.  While Vale
linters also support comment directives, these only apply to linters
with scopes besides `raw`, since the `raw` scope operates on text and
Vale comment directives operate on a parsed AST. By changing these to
`remark` linters, we can examine page structure while also allowing
`lint ignore` comment directives.

How it works sections
---------------------

We use a Vale linter to enforce the presence of a "How it works" H2 in
how-to guides, making sure that these guides begin with an architectural
overview so users can form a mental model of the procedure before they
begin to follow it. (Otherwise, how-to guides read as a plunge into
the unknown.)

Step numbering
--------------

The step numbering linter prevents mistakes in numbering "Step" sections
in how-to guides. These mistakes can happen when, for example, an author
adds or deletes a step after writing a first draft but forgets to update
the numbering in the remaining steps.