Emit a build error on duplicate redirect sources
Duplicate redirect sources in Docusaurus do not interrupt a docs site
build. However, they can have unpredictable results, and cause
Docusaurus to print a warning message that can add clutter to build
logs.

Edit the `getRedirects` function that we use to fetch redirects when
evaluating the Docusaurus configuration file to throw an error if two
redirects have the same source. This function already validates each
redirect, so this change expands the validation logic to validate the
redirect array as a whole.