Build the docs based on release tags
Closes #176

In the `yarn build` script, run another script, `yarn checkout-tags`,
that:
- Queries GitHub for the latest release belonging to each supported
  version.
- Checks out the tag belonging to the latest release in each git
  submodule (the `content/<version>.x` directories).
- Edits the docs site `config.json` file to indicate the release number
  for each branch so we can display it on the version switcher.

This change runs a script on build because it is not possible to
configure `.gitmodules` to check out tags within a branch, only entire
branches.

The script does not exit with an error if we can't find a release, since
the submodule will already be at the latest ref within a major version.
This lets us include the latest docs commits in the `master` branch.

For this approach to work, this change edits the schema of the docs site
`config.json` to add a `release` field, which it uses to set a display
label for the version switcher. To simplify the configuration rewrites,
this change removes the `name` field from `config.json`, since this
field is no longer used.

The script writes the supported versions out of order to `config.json`,
so this change also sorts the versions by name, i.e., version number,
when loading the config.