Respond to taraspos feedback
Do not include a .gitmodules configuration in the docs-website project.
Instead, include a .gitmodules.example file that users can copy for
local development. In production builds, only fetch releases from GitHub
based on `content` directory names.

One consequence of this approach is that we can no longer include the
unreleased major version within the docs site, so this change removes it
from `config.json` and `.gitmodules`.