Edit the release-fetching script
Exit with an error code if a content directory does not have the
expected name format or if the script cannot fetch the latest release
for a major version, since in these cases, there is no recourse for
building the corresponding version of the docs.

Print the response to the curl request for releases for debugging.