Reconfigure remarkVersionAlias in the docs config
The current configuration is based on a misleading parameter name in
`remark-version-alias`, in which the parameter, `latestVersion`,
indicates the version of the docs in the `docs` directory. The current
`docusaurus.config.ts` populates that with the result of
`getLatestVersion`.

However, in Docusaurus parlance, the "latest" version is the default
version of the docs site, while the "current" version is the edge
version.

This change renames the misleading parameter in `remark-verison-alias`
and sets it to the expected value in `docusaurus.config.ts`.