docs: make embedded video by ID rather than by URL
The embedded YouTube video feature is currently implemented in a way that requires us to in put the full url like https://www.youtube.com/watch?v=JJftMAwMld8. Given that there are at least 3 different youtube URL formats, we don't want any obsurity on which to use. This PR updates it so we only need to pass the video ID as an argument and not the correct URL.