Add GHA step to validate config.json
Closes #402

Ensure no `repo_path` or `branch` overrides in config.json make it to
production.

Add the checks in a script to the `scripts` directory, rather than in
the code for loading config.json in `server/config-site`, so we don't
need to branch on whether the code is running in production. Require
that no version config includes `repo_path`, and that `branch` follows
the expected pattern. Include this check in the `test.yaml` workflow.

Remove unnecessary deprecated branches from config.json. Some branch
names don't follow the expected pattern, and in general there is no
reason not to clean these up, as the list will only expand.