Revert Node version change
The change in #437 added v22 as the expected Node version in
package.json. However, GHA workflows that install dependencies in
`gravitational/teleport` turn out not to use this field to choose the
version of Node to install, creating a version mismatch between the GHA
runner and the expected version. Since we only need v22 for the
linter change that we reverted in #437, restore v18 as the expected Node
version.