Allow label-only `DocCardList`s (#460)
Currently, the `DocCardList` component displays a list of items in which
each item contains the label and description of each page in a sidebar
section.

This change adds the possibility of including a `kind` prop in a
`DocCardList` that changes its formatting. The only supported value,
`labelOnly`, only lists the label of each guide. This is helpful in
situations in which descriptions are repetitive and don't provide useful
information on an index page.

If no `kind` is provided, the `DocCardList` lists labels and
descriptions as usual.

Another possibility would be to make this prop a `labelOnly` Boolean,
but it seems likely that we'll add other formatting options in the
future.