Throw an error on an invalid Admonition type (#284)
The default Docusaurus behavior when rendering an Admonition is to print
a warning on an invalid type. This not only causes an unintended result
to render on the docs site, it creates clutter when reading docs build
output. Since Docusaurus build logs are not structured consistently, it
is important to keep it clean for a human reader.