Add `ChangelogLink` (#493)
The `ChangelogLink` component displays a link to the edge version of the
changelog if the user is not currently viewing that version. Since we
are moving to a process in which the changelogs on
`gravitational/teleport` release branches only includes release notes
for their corresponding major versions, this makes it clear for users
where they can find all release notes.