Allow for Boolean `enterprise` frontmatter values
The `enterprise` frontmatter field currently has a string value, and
uses the string as-is when rendering the exclusivity banner. This makes
it difficult to add this frontmatter field in bulk, since we need to
determine a value for this field for each page in the docs that requires
it.

This change allows us to use either a Boolean or a string for the
`enterprise` frontmatter field. If the value is a Boolean, the
exclusivity banner displays "This capability" for the feature name.